requirements.txt
fastapi
filelock
httpx>=0.27.0
line-bot-sdk
uvicorn[standard]
python-dotenv